Reorganize health assets
Renamed the following classes:
  GVFSEnlistmentHealthCalculator -> EnlistmentHealthCalculator
  GVFSEnlistmentPathData -> EnlistmentPathData
  GVFSEnlistmentHealthData -> EnlistmentHealthData
Extracted EnlistmentPathData and EnlistmentHealthData to be standalone
classes instead of inner classes of EnlistmentHealthCalculator.
Created a GVFS.Common.HealthCalculator to house the 3 health classes
